Minor changes - renamed some variables/files
