+++ src/applications/snake_game.py
