+++ src/asset_management/assets.py
@@ -2,24 +2,24 @@ from enum import StrEnum
 
 class DesktopAssets(StrEnum):
     NO_INTERNET_DARK = "desktop/no_internet_icon_dark.png"
-    NO_INTERNET_LIGHT = "desktop/no_internet_icon.png"
+    NO_INTERNET_LIGHT = "desktop/no_internet_icon_light.png"
     PYBROWSE_DARK = "desktop/pybrowse_dark.png"
-    PYBROWSE_LIGHT = "desktop/pybrowse.png"
+    PYBROWSE_LIGHT = "desktop/pybrowse_light.png"
     START_LOGO_DARK = "desktop/python_icon_dark.png"
-    START_LOGO_LIGHT = "desktop/python_icon.png"
+    START_LOGO_LIGHT = "desktop/python_icon_light.png"
     BACKGROUND_LOGO = "desktop/python_logo.png"
     SNAKE_GAME_DARK = "desktop/snake_blue_icon.png"
     SNAKE_GAME_LIGHT = "desktop/snake_yellow_icon.png"
     WIFI_ICON_DARK = "desktop/wifi_icon_dark.png"
-    WIFI_ICON_LIGHT = "desktop/wifi_icon.png"
+    WIFI_ICON_LIGHT = "desktop/wifi_icon_light.png"
     TEXT_FILE_ICON = "desktop/python_logo.png"
 
 
 class PyBrowseAssets(StrEnum):
-    PYBROWSE_GAME_SNAKE_DARK = "pybrowse/anaconda_blue.png"
-    PYBROWSE_GAME_SNAKE_LIGHT = "pybrowse/anaconda_yellow.png"
+    PYBROWSE_GAME_SNAKE_DARK = "pybrowse/snake_blue.png"
+    PYBROWSE_GAME_SNAKE_LIGHT = "pybrowse/snake_yellow.png"
     PYBROWSE_ICON_DARK = "pybrowse/pybrowse_dark.ico"
-    PYBROWSE_ICON_LIGHT = "pybrowse/pybrowse.ico"
+    PYBROWSE_ICON_LIGHT = "pybrowse/pybrowse_light.ico"
 
 
 class SnakeGameAssets(StrEnum):
